#15. Техническое задание проекта спринта №22. Delivery, Order, Payment

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/dto/BookedProductsDto.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,10 +15,10 @@ import lombok.NoArgsConstructor;
 public class BookedProductsDto {
 
     @NotNull
-    private Double deliveryWeight;
+    private BigDecimal deliveryWeight;
 
     @NotNull
-    private Double deliveryVolume;
+    private BigDecimal deliveryVolume;
 
     @NotNull
     private Boolean fragile;
